% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MasturinaPackage.R
\name{new_data}
\alias{new_data}
\title{Replace Missing Values with Column Medians}
\usage{
new_data(data)
}
\arguments{
\item{data}{A data frame. Numeric columns are analyzed for missing values.}
}
\value{
A data frame where NA values in numeric columns are replaced with the column medians.
}
\description{
This function replaces missing values (NA) in numeric columns of a data frame
with the median of the respective column.
}
\examples{
# Example data frame
df <- data.frame(a = c(10, 20, NA, 40),
                 b = c(NA, 50, 60, 70),
                 c = c(50, 60, NA, 80)
                 )
new_data(df)
}
